Extend configuration and sub config.
Use an unique name for sub type name. The original assumption is that
all type is the same type so they share the same default vaule. With the
support of syncDefaultValueToCurrent, this assumption doesn't hold
anymore. So we need to individually assign every option with a unique
type name. Here we use a spell character '$' as separator, and the name
will be named as 'Path$To$Value$TypeName'.