Further workaround for wengxt/gnome-shell-extension-kimpanel#56
Delay the SNI registration to flip-flap on using sni or not.

If we register and close dbus connection very quickly, it is likekly
that client is not able to capture the destruction of our SNI. (Still
could be a bug in the client, but whatever). Delay the registration, so
it can save us some message for kimpanel + SNI startup.