Stop key repeat timer on wayland keyboard grab cleared (#271)
This should fix one of the unexpected key inputs problem on sway:
https://github.com/swaywm/sway/issues/6193

The problem occurs as follows:

 1. press Ctrl+d to exit from terminal emulator
 2. fcitx captures the key press event and enables repeat timer
 3. terminal emulator exits (while the key is still pressed)
 4. fcitx sends key repeat events indefinitely (since there's no release event)

It's unclear whether this is a bug of the compositor (i.e. sway) or not, but
I think it's better for fcitx to not carry the key-repeat state over to new
keyboard grab.