connectableobject: introduce FCITX_DECLARE_SIGNAL_WITH_COMBINER
In order to correctly downcast SignalBase, in the next step,
we will switch all initialisation of SignalAdaptor to use
SignalType::combinerType.

Let's introduce a new macro to ease to initialisation of those
signals which employs custom combiner.

Arguably, the custom combiner is an implementation detail,
thus it'll be put in source files. Hence, the macro is intended
to be used outside of class declaration.