Always call setOverlayPanel, in order to trigger show() in kwin.
Right now it is not clear if how "should" requestHidePanel from
text-input-v2 to be used with zwp input method v1. It's is now purely in a
implementation defined zone. Luckly, there is no other implementation
we need to care about. Just use this trick to ensure if input
method request to show it, it is always visible.