Make weston happy with our input panel implementation.
Due to the fact that there is only weston/kwin as implementation. Weston
does not like us to call set_overlay_panel multiple times. Luckily, it
is OK to get multiple input panel surface for both kwin/weston. There is
an additional limitation for weston, that one surface can be only
requested to be input panel once, even input panel surface is destroyed.

Thus, the implementation is changed to following:
1. upon hide the window, destory all surface object. This exposes an
   issue that destroyed surface may still emit leave signal, need an
   additional check for that.
2. upon showing the window, request surface keep it until hide. This may
   also workaround qt's request hide but not show again issue.

Fix #430