Add a new variant of ProcessKeyEvent to allow piggy back event happened in the middle of key process.
Prepare for #445.

In hangul, the event order is desired to be arrive in the order of
1. commit string
2. handle event

Due to the gtk4 can only to sync event handling properly for now, we
need a way to handle commit string before the key event handle reply.
Thus we extend ProcessKeyEvent into ProcessKeyEventBatch, to make it
possible to piggy back event into the same reply. This has similar
effect like KeyEventOrderFix flag.