Add virtual key handling to InputContext and WaylandIMInputContextV1
The current InputContext has the `keyEvent` method, but what it can do
is limited.
In WaylandIMInputContextV1, key inputs are processed in the `keyCallback`
method, but there is no way to pass virtual key inputs to it.

This fix adds new methods `virtualKeyEvent` and `virtualKeyEventImpl`
to InputContext.
By default, it works the same as the existing `keyEvent` method.
But `virtualKeyEventImpl` can be overridden by each implementation
to define the same behavior as physical keys.