Fix spell with "-" word
In fcitx4 actually we separate the candidate with two field
(display, commit). And we didn't correctly return prefix+hint for custom
dictionary.

This fixes this, and also, since we handles the upper/lowercases in the dict level,
the engine level case-fix does not make sense anymore. So push down to
the dict. Especially we already has case fix one-run in the custom dict.
Doesn't really make sense if we do that twice.