Implement NotifyFocusOut signal for #687
In fcitx, the focus group mechanism may kick other input context out of
focus. When this happens, the client side unfocus commit won't be
triggered and the preedit will be simply reset.

This mainly only happens for dbus based client because of the
inconsistent state between dbus/display protocol.

To avoid such issue, introduce a new NotifyFocusOut signal when focus out
actually happens in fcitx to dbus. The input context may listen to it
and do client side preedit commit.