Reduce the requirements of unnecessary key event forwarding.
In a previous fix of keyboard engine, update preedit to empty is commonly done
to ensure there is nothing in the preedit. This introduces some empty
update preedit for keyboard engine and uses the code path for key event
order fix. This makes almost all key event need to be re-forwarded,
which is unnecssary.

Add a new helper function to check if the strict order matters:
1. all pending events are update preedit.
2. preedit is empty.

If above condition satisfies and we don't need to need forward the key,
allow it to be handled out of order.

Attempt Fix #702