Make normalize() convert super2 into super
Come to think about it, it is actually more erroneous if we keep a new
state here and there. Since we already mess with the state with
normalize, just convert Super2 into our own super and call it a day.

This saves our life for other Key member functions since most of them
expect the input to be normalized.

Fix #747