Fix icon name mapping in flatpak
Now we have a much of icons that uses "fcitx_" as the name, including
mozc, skk, kkc, rime, to avoid xdg icon spec based fallback.

But we missed the flatpak naming mapping for those icons. Update the
condition here, and also handle the special name "fcitx" (I don't think
we do have code that uses "fcitx" as icon, but well.. )

https://github.com/fcitx/mozc/issues/53#issuecomment-2105739015