The virtual keyboard should show or hide automatically if enabled (#1086)
1. When the user interacts with some applications rather than the
virtual keyboard itself, the virtual keyboard should show or hide
if it is enabled.

2. For example, if the user clicks a window and the window has a
text input control which has the focus and accepts the input from
the keyboard, then the virtual keyboard should show automatically
if the auto show switch is enabled.

3. Then, if the window loses the focus, the virtual keyboard should
hide automatically if the auto hide switch is enabled.