Add new global option to allow change the time limit on modifier only key triggering
This is intended to solve some issue when user have other intentions.

For example, user may press "Shift" to do some text selection with
mouse. However, input method may not be capable (theoratically
possible on certain platform, but it's not implemented and not possible
on all platform) to detect user actually want to use mouse. If application
does not send "reset()", input method may wrongly trigger the action.

Some other cases include:
- Ctrl and scroll mouse
- user wants to type Shift+A, but regrets after press shift. So user
  release all key but accidentally triggers "Shift" to switch input
  method.

This is not perfect but should help solve some unintended action.