Use _NET_WM_WINDOW_TYPE_COMBO as input window type under X11.
Pantheon has a bug that can't not display popup window under Xwayland
twice. While the bug is reported, it's not clear when it may be fixed.
Also, input method window is indeed more like a combo box and wmspec
mentioned that it's also a common override redirect. So try to use that
instead.

Workaround #1209