Only set layout to kde config if zwp_input_method_v1 is available a.k.a it is the chosen input method under virtual keyboard. (#1280)
Recently there was a misconfigured dependency on Arch in AUR's sdl package,
that may accidentually make AUR user poll in fcitx onto their system.
Which causes a lot of confusion for KDE users.

Since we now have selfDiagnose() on wayland, it is actually asking user
to be aware they're configuring fcitx in kde's system settings. If user
does not follow the guide, we'd assume user may not want to use fcitx to
avoid accidentally override user's keyboard config.

On arch there was also fcitx5's xdg autostart enabled, so it's very
likely if user installed it will be run by default.