Always prefer the first langauge in layout instead of find an alpha 2 code.
It has been confusing that tibetan is shown as Chinese since we prefer
alpha 2 code in a latter field. Usually, the first is the most accurate
language code.

Since we will make configtool to always use iso639 language, it
shouldn't be a problem that alpha3 is used when there's no alpha 2.

fcitx/fcitx5-configtool#92